+++ sunpy/map/mapbase.py
@@ -4,46 +4,40 @@ Map is a generic Map class from which all other Map classes inherit from.
 from __future__ import absolute_import, division, print_function
 from sunpy.extern.six.moves import range
 
-#pylint: disable=E1101,E1121,W0404,W0613
-__authors__ = ["Russell Hewett, Stuart Mumford, Keith Hughitt, Steven Christe", "Jack Ireland"]
-__email__ = "stuart@mumford.me.uk"
-
 import warnings
 import inspect
 from abc import ABCMeta
 from copy import deepcopy
-from collections import OrderedDict
+from collections import OrderedDict, namedtuple
 
 import numpy as np
 import matplotlib.pyplot as plt
 from matplotlib import patches, cm, colors
 
 import astropy.wcs
+import astropy.units as u
 from astropy.coordinates import Longitude, Latitude
-
-from sunpy.image.transform import affine_transform
-from .nddata_compat import NDDataCompat as NDData
 
 import sunpy.io as io
 import sunpy.wcs as wcs
 import sunpy.coordinates
+from sunpy import config
+from sunpy.extern import six
 from sunpy.visualization import toggle_pylab, wcsaxes_compat
 from sunpy.sun import constants
 from sunpy.sun import sun
 from sunpy.time import parse_time, is_time
+from sunpy.image.transform import affine_transform
 from sunpy.image.rescale import reshape_image_to_4d_superpixel
 from sunpy.image.rescale import resample as sunpy_image_resample
 
-from sunpy.extern import six
+from astropy.nddata import NDData
 
-import astropy.units as u
-
-from collections import namedtuple
+TIME_FORMAT = config.get("general", "time_format")
 Pair = namedtuple('Pair', 'x y')
 
-
-from sunpy import config
-TIME_FORMAT = config.get("general", "time_format")
+__authors__ = "Russell Hewett, Stuart Mumford, Keith Hughitt," \
+              " Steven Christe, Jack Ireland"
 
 __all__ = ['GenericMap']
 
@@ -127,21 +121,14 @@ class GenericMap(NDData):
     Pair(x=Unit("arcsec"), y=Unit("arcsec"))
     >>> aia.peek()   # doctest: +SKIP
 
-    References
-    ----------
-    | http://docs.scipy.org/doc/numpy/reference/arrays.classes.html
-    | http://docs.scipy.org/doc/numpy/user/basics.subclassing.html
-    | http://docs.scipy.org/doc/numpy/reference/ufuncs.html
-    | http://www.scipy.org/Subclasses
-
     Notes
     -----
 
     A number of the properties of this class are returned as two-value named
-    tuples that can either be indexed by position ([0] or [1]) or be accessed by
-    name (.x or .y).  The names "x" and "y" here refer to the first and second
-    axes of the map, and may not necessarily correspond to any similarly named
-    axes in the coordinate system.
+    tuples that can either be indexed by position ([0] or [1]) or be accessed
+    by name (.x or .y). The names "x" and "y" here refer to the first and
+    second axes of the map, and may not necessarily correspond to any similarly
+    named axes in the coordinate system.
 
     This class makes some assumptions about the WCS information contained in
     the meta data. The first and most extensive assumption is that it is
@@ -154,12 +141,14 @@ class GenericMap(NDData):
       notation using equations 32 in Thompson (2006).
 
     * If a CDi_j matrix is provided it is assumed that it can be converted to a
-      PCi_j matrix and CDELT keywords as described in Greisen & Calabretta (2002).
+      PCi_j matrix and CDELT keywords as described in
+      `Greisen & Calabretta (2002) <http://dx.doi.org/10.1051/0004-6361:20021327>`_
 
     * The 'standard' FITS keywords that are used by this class are the PCi_j
-      matrix and CDELT, along with the other keywords specified in the WCS papers.
-      All subclasses of this class must convert their header information to
-      this formalism. The CROTA to PCi_j conversion is done in this class.
+      matrix and CDELT, along with the other keywords specified in the WCS
+      papers. All subclasses of this class must convert their header
+      information to this formalism. The CROTA to PCi_j conversion is done in
+      this class.
 
     .. warning::
         This class currently assumes that a header with the CDi_j matrix
@@ -169,7 +158,7 @@ class GenericMap(NDData):
         keywords are exactly equal.
     """
 
-    def __init__(self, data, header, **kwargs):
+    def __init__(self, data, header, plot_settings=None, **kwargs):
 
         super(GenericMap, self).__init__(data, meta=header, **kwargs)
 
@@ -189,17 +178,21 @@ class GenericMap(NDData):
             norm = None
         else:
             norm = colors.Normalize()
+
         # Visualization attributes
         self.plot_settings = {'cmap': cm.gray,
                               'norm': norm,
                               'interpolation': 'nearest',
                               'origin': 'lower'
                               }
+        if plot_settings:
+            self.plot_settings.update(plot_settings)
 
     def __getitem__(self, key):
         """ This should allow indexing by physical coordinate """
         raise NotImplementedError(
-            "The ability to index Map by physical coordinate is not yet implemented.")
+            "The ability to index Map by physical"
+            " coordinate is not yet implemented.")
 
     def __repr__(self):
         if not self.observatory:
@@ -219,11 +212,19 @@ scale:\t\t {scale}
 
 """.format(dtype=self.__class__.__name__,
            obs=self.observatory, inst=self.instrument, det=self.detector,
-           meas=self.measurement, wave=self.wavelength, date=self.date, dt=self.exposure_time,
+           meas=self.measurement, wave=self.wavelength, date=self.date,
+           dt=self.exposure_time,
            dim=u.Quantity(self.dimensions),
            scale=u.Quantity(self.scale),
-           tmf=TIME_FORMAT)
-+ self.data.__repr__())
+           tmf=TIME_FORMAT) + self.data.__repr__())
+
+    @classmethod
+    def _new_instance(cls, data, meta, plot_settings=None, **kwargs):
+        """
+        Instantiate a new instance of this class using given data.
+        This is a shortcut for ``type(self)(data, meta, plot_settings)``
+        """
+        return cls(data, meta, plot_settings=plot_settings, **kwargs)
 
     @property
     def wcs(self):
@@ -383,16 +384,11 @@ scale:\t\t {scale}
         """Image observation time"""
         time = parse_time(self.meta.get('date-obs', 'now'))
         if time is None:
-            warnings.warn_explicit("Missing metadata for observation time. Using current time.",
-                                       Warning, __file__, inspect.currentframe().f_back.f_lineno)
+            warnings.warn_explicit("Missing metadata for observation time."
+                                   " Using current time.",
+                                   Warning, __file__,
+                                   inspect.currentframe().f_back.f_lineno)
         return parse_time(time)
-
-#    @date.setter
-#    def date(self, new_date):
-#        self.meta['date-obs'] = new_date
-#        #propagate change to malformed FITS keywords
-#        if is_time(self.meta.get('date_obs', None)):
-#            self.meta['date_obs'] = new_date
 
     @property
     def detector(self):
@@ -405,8 +401,10 @@ scale:\t\t {scale}
         dsun = self.meta.get('dsun_obs', None)
 
         if dsun is None:
-            warnings.warn_explicit("Missing metadata for Sun-spacecraft separation: assuming Sun-Earth distance",
-                                   Warning, __file__, inspect.currentframe().f_back.f_lineno)
+            warnings.warn_explicit("Missing metadata for Sun-spacecraft"
+                                   " separation: assuming Sun-Earth distance",
+                                   Warning, __file__,
+                                   inspect.currentframe().f_back.f_lineno)
             dsun = sun.sunearth_distance(self.date).to(u.m)
 
         return u.Quantity(dsun, 'm')
@@ -424,22 +422,25 @@ scale:\t\t {scale}
     @property
     def measurement(self):
         """Measurement name, defaults to the wavelength of image"""
-        return u.Quantity(self.meta.get('wavelnth', 0), self.meta.get('waveunit', ""))
+        return u.Quantity(self.meta.get('wavelnth', 0),
+                          self.meta.get('waveunit', ""))
 
     @property
     def wavelength(self):
         """wavelength of the observation"""
-        return u.Quantity(self.meta.get('wavelnth', 0), self.meta.get('waveunit', ""))
+        return u.Quantity(self.meta.get('wavelnth', 0),
+                          self.meta.get('waveunit', ""))
 
     @property
     def observatory(self):
         """Observatory or Telescope name"""
-        return self.meta.get('obsrvtry', self.meta.get('telescop', "")).replace("_", " ")
+        return self.meta.get('obsrvtry',
+                             self.meta.get('telescop', "")).replace("_", " ")
 
     @property
     def xrange(self):
         """Return the X range of the image from edge to edge."""
-        #TODO: This should be reading from the WCS object
+        # TODO: This should be reading from the WCS object
         xmin = self.center.x - self.dimensions[0] / 2. * self.scale.x
         xmax = self.center.x + self.dimensions[0] / 2. * self.scale.x
         return u.Quantity([xmin, xmax])
@@ -447,20 +448,18 @@ scale:\t\t {scale}
     @property
     def yrange(self):
         """Return the Y range of the image from edge to edge."""
-        #TODO: This should be reading from the WCS object
+        # TODO: This should be reading from the WCS object
         ymin = self.center.y - self.dimensions[1] / 2. * self.scale.y
         ymax = self.center.y + self.dimensions[1] / 2. * self.scale.y
         return u.Quantity([ymin, ymax])
 
     @property
     def center(self):
-        """The offset between the center of the Sun and the center of the map."""
-        return Pair(wcs.get_center(self.dimensions[0], self.scale.x,
-                                   self.reference_pixel.x,
-                                   self.reference_coordinate.x),
-                    wcs.get_center(self.dimensions[1], self.scale.y,
-                                   self.reference_pixel.y,
-                                   self.reference_coordinate.y))
+        """
+        Return the world (data) coordinates of the center pixel of the array.
+        """
+        center = u.Quantity(self.dimensions) / 2.
+        return Pair(*self.pixel_to_data(*center))
 
     @property
     def shifted_value(self):
@@ -470,9 +469,11 @@ scale:\t\t {scale}
 
     @u.quantity_input(x=u.deg, y=u.deg)
     def shift(self, x, y):
-        """Returns a map shifted by a specified amount to, for example, correct for a bad
-        map location. These values are applied directly to the `~sunpy.map.GenericMap.reference_coordinate`.
-        To check how much shift has already been applied see `~sunpy.map.GenericMap.shifted_value`
+        """
+        Returns a map shifted by a specified amount to, for example, correct
+        for a bad map location. These values are applied directly to the
+        `~sunpy.map.GenericMap.reference_coordinate`. To check how much shift
+        has already been applied see `~sunpy.map.GenericMap.shifted_value`
 
         Parameters
         ----------
@@ -488,16 +489,19 @@ scale:\t\t {scale}
         out : `~sunpy.map.GenericMap` or subclass
             A new shifted Map.
         """
-        new_map = deepcopy(self)
-        new_map._shift = Pair(self.shifted_value.x + x, self.shifted_value.y + y)
-
         new_meta = self.meta.copy()
 
         # Update crvals
-        new_meta['crval1'] = ((self.meta['crval1'] * self.spatial_units.x + x).to(self.spatial_units.x)).value
-        new_meta['crval2'] = ((self.meta['crval2'] * self.spatial_units.y + y).to(self.spatial_units.y)).value
+        new_meta['crval1'] = ((self.meta['crval1'] *
+                               self.spatial_units.x + x).to(self.spatial_units.x)).value
+        new_meta['crval2'] = ((self.meta['crval2'] *
+                               self.spatial_units.y + y).to(self.spatial_units.y)).value
 
-        new_map.meta = new_meta
+        #Create new map with the modification
+        new_map = self._new_instance(self.data, new_meta, self.plot_settings)
+
+        new_map._shift = Pair(self.shifted_value.x + x,
+                              self.shifted_value.y + y)
 
         return new_map
 
@@ -511,11 +515,14 @@ scale:\t\t {scale}
         """Radius of the Sun."""
         rsun_arcseconds = self.meta.get('rsun_obs',
                                         self.meta.get('solar_r',
-                                                      self.meta.get('radius', None)))
+                                                      self.meta.get('radius',
+                                                                    None)))
 
         if rsun_arcseconds is None:
-            warnings.warn_explicit("Missing metadata for solar radius: assuming photospheric limb as seen from Earth",
-                                   Warning, __file__, inspect.currentframe().f_back.f_lineno)
+            warnings.warn_explicit("Missing metadata for solar radius:"
+                                   " assuming photospheric limb as seen from Earth",
+                                   Warning, __file__,
+                                   inspect.currentframe().f_back.f_lineno)
             rsun_arcseconds = sun.solar_semidiameter_angular_size(self.date).to('arcsec').value
 
         return u.Quantity(rsun_arcseconds, 'arcsec')
@@ -532,8 +539,10 @@ scale:\t\t {scale}
         carrington_longitude = self.meta.get('crln_obs', None)
 
         if carrington_longitude is None:
-            warnings.warn_explicit("Missing metadata for Carrington longitude: assuming Earth-based observer",
-                                   Warning, __file__, inspect.currentframe().f_back.f_lineno)
+            warnings.warn_explicit("Missing metadata for Carrington longitude:"
+                                   " assuming Earth-based observer",
+                                   Warning, __file__,
+                                   inspect.currentframe().f_back.f_lineno)
             carrington_longitude = (sun.heliographic_solar_center(self.date))[0]
 
         return u.Quantity(carrington_longitude, 'deg')
@@ -546,8 +555,10 @@ scale:\t\t {scale}
                                                             self.meta.get('solar_b0', None)))
 
         if heliographic_latitude is None:
-            warnings.warn_explicit("Missing metadata for heliographic latitude: assuming Earth-based observer",
-                                   Warning, __file__, inspect.currentframe().f_back.f_lineno)
+            warnings.warn_explicit("Missing metadata for heliographic latitude:"
+                                   " assuming Earth-based observer",
+                                   Warning, __file__,
+                                   inspect.currentframe().f_back.f_lineno)
             heliographic_latitude = (sun.heliographic_solar_center(self.date))[1]
 
         return u.Quantity(heliographic_latitude, 'deg')
@@ -567,26 +578,34 @@ scale:\t\t {scale}
     @property
     def reference_pixel(self):
         """Reference point axes in pixels (i.e. crpix1, crpix2)"""
-        return Pair(self.meta.get('crpix1', (self.meta.get('naxis1') + 1) / 2.) * u.pixel,
-                    self.meta.get('crpix2', (self.meta.get('naxis2') + 1) / 2.) * u.pixel)
+        return Pair(self.meta.get('crpix1',
+                                  (self.meta.get('naxis1') + 1) / 2.) * u.pixel,
+                    self.meta.get('crpix2',
+                                  (self.meta.get('naxis2') + 1) / 2.) * u.pixel)
 
     @property
     def scale(self):
-        """Image scale along the x and y axes in units/pixel (i.e. cdelt1, cdelt2)"""
-        #TODO: Fix this if only CDi_j matrix is provided
+        """
+        Image scale along the x and y axes in units/pixel (i.e. cdelt1, cdelt2)
+        """
+        # TODO: Fix this if only CDi_j matrix is provided
         return Pair(self.meta.get('cdelt1', 1.) * self.spatial_units.x / u.pixel,
                     self.meta.get('cdelt2', 1.) * self.spatial_units.y / u.pixel)
 
     @property
     def spatial_units(self):
-        """Image coordinate units along the x and y axes (i.e. cunit1, cunit2)."""
+        """
+        Image coordinate units along the x and y axes (i.e. cunit1, cunit2).
+        """
         return Pair(u.Unit(self.meta.get('cunit1', 'arcsec')),
                     u.Unit(self.meta.get('cunit2', 'arcsec')))
 
     @property
     def rotation_matrix(self):
-        """Matrix describing the rotation required to align solar North with
-        the top of the image."""
+        """
+        Matrix describing the rotation required to align solar North with
+        the top of the image.
+        """
         if 'PC1_1' in self.meta:
             return np.matrix([[self.meta['PC1_1'], self.meta['PC1_2']],
                               [self.meta['PC2_1'], self.meta['PC2_2']]])
@@ -618,8 +637,8 @@ scale:\t\t {scale}
 # #### Miscellaneous #### #
 
     def _fix_date(self):
-        # Check commonly used but non-standard FITS keyword for observation time
-        # and correct the keyword if we can.  Keep updating old one for
+        # Check commonly used but non-standard FITS keyword for observation
+        # time and correct the keyword if we can. Keep updating old one for
         # backwards compatibility.
         if is_time(self.meta.get('date_obs', None)):
             self.meta['date-obs'] = self.meta['date_obs']
@@ -648,7 +667,8 @@ scale:\t\t {scale}
 
     def _get_cmap_name(self):
         """Build the default color map name."""
-        cmap_string = self.observatory + self.meta['detector'] + str(int(self.wavelength.to('angstrom').value))
+        cmap_string = (self.observatory + self.meta['detector'] +
+                       str(int(self.wavelength.to('angstrom').value)))
         return cmap_string.lower()
 
     def _validate_meta(self):
@@ -670,7 +690,8 @@ scale:\t\t {scale}
                 u.Unit(self.meta.get(meta_property),
                        parse_strict='silent').physical_type == 'unknown'):
 
-                warnings.warn("Unknown value for "+meta_property.upper(), Warning)
+                warnings.warn("Unknown value for " + meta_property.upper(),
+                              Warning)
 
 
 # #### Data conversion routines #### #
@@ -704,7 +725,8 @@ scale:\t\t {scale}
         y : `~astropy.units.Quantity`
             Pixel coordinate on the CTYPE2 axis.
         """
-        x, y = self.wcs.wcs_world2pix(x.to(u.deg).value, y.to(u.deg).value, origin)
+        x, y = self.wcs.wcs_world2pix(x.to(u.deg).value,
+                                      y.to(u.deg).value, origin)
 
         return x * u.pixel, y * u.pixel
 
@@ -805,7 +827,8 @@ scale:\t\t {scale}
 
         References
         ----------
-        * `Rebinning <http://www.scipy.org/Cookbook/Rebinning>`_ (Original source, 2011/11/19)
+        * `Rebinning <http://www.scipy.org/Cookbook/Rebinning>`_ (Original
+          source, 2011/11/19)
         """
 
         # Note: because the underlying ndarray is transposed in sense when
@@ -816,13 +839,12 @@ scale:\t\t {scale}
 
         # Make a copy of the original data and perform resample
         new_data = sunpy_image_resample(self.data.copy().T, dimensions,
-                                    method, center=True)
+                                        method, center=True)
         new_data = new_data.T
 
         scale_factor_x = float(self.dimensions[0] / dimensions[0])
         scale_factor_y = float(self.dimensions[1] / dimensions[1])
 
-        new_map = deepcopy(self)
         # Update image scale and number of pixels
         new_meta = self.meta.copy()
 
@@ -840,8 +862,7 @@ scale:\t\t {scale}
         new_meta['crval2'] = self.center.y.value
 
         # Create new map instance
-        new_map.data = new_data
-        new_map.meta = new_meta
+        new_map = self._new_instance(new_data, new_meta, self.plot_settings)
         return new_map
 
     def rotate(self, angle=None, rmatrix=None, order=4, scale=1.0,
@@ -865,13 +886,13 @@ scale:\t\t {scale}
         rmatrix : 2x2
             Linear transformation rotation matrix.
         order : int 0-5
-            Interpolation order to be used. When using scikit-image this parameter
-            is passed into :func:`skimage.transform.warp` (e.g., 4 corresponds to
-            bi-quartic interpolation).
+            Interpolation order to be used. When using scikit-image this
+            parameter is passed into :func:`skimage.transform.warp` (e.g., 4
+            corresponds to bi-quartic interpolation).
             When using scipy it is passed into
-            :func:`scipy.ndimage.interpolation.affine_transform` where it controls
-            the order of the spline.
-            Faster performance may be obtained at the cost of accuracy by using lower values.
+            :func:`scipy.ndimage.interpolation.affine_transform` where it
+            controls the order of the spline. Faster performance may be
+            obtained at the cost of accuracy by using lower values.
             Default: 4
         scale : float
             A scale factor for the image, default is no scaling
@@ -895,7 +916,8 @@ scale:\t\t {scale}
 
         See Also
         --------
-        sunpy.image.transform.affine_transform : The routine this method calls for the rotation.
+        sunpy.image.transform.affine_transform : The routine this method calls
+        for the rotation.
 
         Notes
         -----
@@ -903,8 +925,8 @@ scale:\t\t {scale}
         This function will also convert a CDi_j matrix to a PCi_j matrix.
 
         See :func:`sunpy.image.transform.affine_transform` for details on the
-        transformations, situations when the underlying data is modified prior to rotation,
-        and differences from IDL's rot().
+        transformations, situations when the underlying data is modified prior
+        to rotation, and differences from IDL's rot().
         """
         if angle is not None and rmatrix is not None:
             raise ValueError("You cannot specify both an angle and a matrix")
@@ -921,7 +943,7 @@ scale:\t\t {scale}
                     raise u.UnitsError("Argument '{0}' to function '{1}'"
                                        " must be in units convertable to"
                                        " '{2}'.".format('angle', 'rotate',
-                                                      u.deg.to_string()))
+                                                        u.deg.to_string()))
 
             # Either there is no .unit or no .is_equivalent
             except AttributeError:
@@ -930,8 +952,8 @@ scale:\t\t {scale}
                 else:
                     error_msg = "no 'unit' attribute"
                 raise TypeError("Argument '{0}' to function '{1}' has {2}. "
-                      "You may want to pass in an astropy Quantity instead."
-                         .format('angle', 'rotate', error_msg))
+                                "You may want to pass in an astropy Quantity instead."
+                                 .format('angle', 'rotate', error_msg))
 
         # Interpolation parameter sanity
         if order not in range(6):
@@ -942,9 +964,8 @@ scale:\t\t {scale}
         rotation_center = u.Quantity([self.reference_coordinate.x,
                                       self.reference_coordinate.y])
 
-        # Copy Map
-        new_map = deepcopy(self)
-
+        # Copy meta data
+        new_meta = self.meta.copy()
         if angle is not None:
             # Calculate the parameters for the affine_transform
             c = np.cos(np.deg2rad(angle))
@@ -952,39 +973,40 @@ scale:\t\t {scale}
             rmatrix = np.matrix([[c, -s], [s, c]])
 
         # Calculate the shape in pixels to contain all of the image data
-        extent = np.max(np.abs(np.vstack((new_map.data.shape * rmatrix,
-                                          new_map.data.shape * rmatrix.T))), axis=0)
+        extent = np.max(np.abs(np.vstack((self.data.shape * rmatrix,
+                                          self.data.shape * rmatrix.T))), axis=0)
         # Calculate the needed padding or unpadding
-        diff = np.asarray(np.ceil((extent - new_map.data.shape) / 2), dtype=int).ravel()
+        diff = np.asarray(np.ceil((extent - self.data.shape) / 2), dtype=int).ravel()
         # Pad the image array
         pad_x = int(np.max((diff[1], 0)))
         pad_y = int(np.max((diff[0], 0)))
-        new_map.data = np.pad(new_map.data,
+
+        new_data = np.pad(self.data,
                               ((pad_y, pad_y), (pad_x, pad_x)),
                               mode='constant',
                               constant_values=(missing, missing))
-        new_map.meta['crpix1'] += pad_x
-        new_map.meta['crpix2'] += pad_y
+        new_meta['crpix1'] += pad_x
+        new_meta['crpix2'] += pad_y
 
         # All of the following pixel calculations use a pixel origin of 0
 
-        pixel_array_center = (np.flipud(new_map.data.shape) - 1) / 2.0
+        pixel_array_center = (np.flipud(new_data.shape) - 1) / 2.0
 
         # Convert the axis of rotation from data coordinates to pixel coordinates
-        pixel_rotation_center = u.Quantity(new_map.data_to_pixel(*rotation_center,
-                                                                 origin=0)).value
+        pixel_rotation_center = u.Quantity(self.data_to_pixel(*rotation_center,
+                                                               origin=0)).value
         if recenter:
             pixel_center = pixel_rotation_center
         else:
             pixel_center = pixel_array_center
 
         # Apply the rotation to the image data
-        new_map.data = affine_transform(new_map.data.T,
-                                        np.asarray(rmatrix),
-                                        order=order, scale=scale,
-                                        image_center=np.flipud(pixel_center),
-                                        recenter=recenter, missing=missing,
-                                        use_scipy=use_scipy).T
+        new_data = affine_transform(new_data.T,
+                                    np.asarray(rmatrix),
+                                    order=order, scale=scale,
+                                    image_center=np.flipud(pixel_center),
+                                    recenter=recenter, missing=missing,
+                                    use_scipy=use_scipy).T
 
         if recenter:
             new_reference_pixel = pixel_array_center
@@ -995,45 +1017,47 @@ scale:\t\t {scale}
             new_reference_pixel = np.array(new_reference_pixel).ravel()
 
         # Define the new reference_pixel
-        new_map.meta['crval1'] = rotation_center[0].value
-        new_map.meta['crval2'] = rotation_center[1].value
-        new_map.meta['crpix1'] = new_reference_pixel[0] + 1 # FITS pixel origin is 1
-        new_map.meta['crpix2'] = new_reference_pixel[1] + 1 # FITS pixel origin is 1
+        new_meta['crval1'] = rotation_center[0].value
+        new_meta['crval2'] = rotation_center[1].value
+        new_meta['crpix1'] = new_reference_pixel[0] + 1 # FITS pixel origin is 1
+        new_meta['crpix2'] = new_reference_pixel[1] + 1 # FITS pixel origin is 1
 
         # Unpad the array if necessary
         unpad_x = -np.min((diff[1], 0))
         if unpad_x > 0:
-            new_map.data = new_map.data[:, unpad_x:-unpad_x]
-            new_map.meta['crpix1'] -= unpad_x
+            new_data = new_data[:, unpad_x:-unpad_x]
+            new_meta['crpix1'] -= unpad_x
         unpad_y = -np.min((diff[0], 0))
         if unpad_y > 0:
-            new_map.data = new_map.data[unpad_y:-unpad_y, :]
-            new_map.meta['crpix2'] -= unpad_y
+            new_data = new_data[unpad_y:-unpad_y, :]
+            new_meta['crpix2'] -= unpad_y
 
         # Calculate the new rotation matrix to store in the header by
         # "subtracting" the rotation matrix used in the rotate from the old one
         # That being calculate the dot product of the old header data with the
         # inverse of the rotation matrix.
-        pc_C = np.dot(new_map.rotation_matrix, rmatrix.I)
-        new_map.meta['PC1_1'] = pc_C[0,0]
-        new_map.meta['PC1_2'] = pc_C[0,1]
-        new_map.meta['PC2_1'] = pc_C[1,0]
-        new_map.meta['PC2_2'] = pc_C[1,1]
+        pc_C = np.dot(self.rotation_matrix, rmatrix.I)
+        new_meta['PC1_1'] = pc_C[0,0]
+        new_meta['PC1_2'] = pc_C[0,1]
+        new_meta['PC2_1'] = pc_C[1,0]
+        new_meta['PC2_2'] = pc_C[1,1]
 
         # Update pixel size if image has been scaled.
         if scale != 1.0:
-            new_map.meta['cdelt1'] = (new_map.scale.x / scale).value
-            new_map.meta['cdelt2'] = (new_map.scale.y / scale).value
+            new_meta['cdelt1'] = (self.scale.x / scale).value
+            new_meta['cdelt2'] = (self.scale.y / scale).value
 
         # Remove old CROTA kwargs because we have saved a new PCi_j matrix.
-        new_map.meta.pop('CROTA1', None)
-        new_map.meta.pop('CROTA2', None)
+        new_meta.pop('CROTA1', None)
+        new_meta.pop('CROTA2', None)
         # Remove CDi_j header
-        new_map.meta.pop('CD1_1', None)
-        new_map.meta.pop('CD1_2', None)
-        new_map.meta.pop('CD2_1', None)
-        new_map.meta.pop('CD2_2', None)
+        new_meta.pop('CD1_1', None)
+        new_meta.pop('CD1_2', None)
+        new_meta.pop('CD2_1', None)
+        new_meta.pop('CD2_2', None)
 
+        #Create new map with the modification
+        new_map = self._new_instance(new_data, new_meta, self.plot_settings)
         return new_map
 
     def submap(self, range_a, range_b):
@@ -1172,17 +1196,20 @@ scale:\t\t {scale}
         new_data = self.data[yslice, xslice].copy()
 
         # Make a copy of the header with updated centering information
-        new_map = deepcopy(self)
-        new_map.meta['crpix1'] = self.reference_pixel.x.value - x_pixels[0]
-        new_map.meta['crpix2'] = self.reference_pixel.y.value - y_pixels[0]
-        new_map.meta['naxis1'] = new_data.shape[1]
-        new_map.meta['naxis2'] = new_data.shape[0]
+        new_meta = self.meta.copy()
+        new_meta['crpix1'] = self.reference_pixel.x.value - x_pixels[0]
+        new_meta['crpix2'] = self.reference_pixel.y.value - y_pixels[0]
+        new_meta['naxis1'] = new_data.shape[1]
+        new_meta['naxis2'] = new_data.shape[0]
 
         # Create new map instance
-        new_map.data = new_data
         if self.mask is not None:
-            new_map.mask = self.mask[yslice, xslice].copy()
-
+            new_mask = self.mask[yslice, xslice].copy()
+            #Create new map with the modification
+            new_map = self._new_instance(new_data, new_meta, self.plot_settings, mask=new_mask)
+            return new_map
+        #Create new map with the modification
+        new_map = self._new_instance(new_data, new_meta, self.plot_settings)
         return new_map
 
     @u.quantity_input(dimensions=u.pixel, offset=u.pixel)
@@ -1240,8 +1267,9 @@ scale:\t\t {scale}
         new_array = func(func(reshaped, axis=3), axis=1)
 
         # Update image scale and number of pixels
-        new_map = deepcopy(self)
-        new_meta = new_map.meta
+
+        # create copy of new meta data
+        new_meta = self.meta.copy()
 
         new_nx = new_array.shape[1]
         new_ny = new_array.shape[0]
@@ -1261,11 +1289,14 @@ scale:\t\t {scale}
 
         # Create new map instance
         if self.mask is not None:
-            new_map.data = np.ma.getdata(new_array)
-            new_map.mask = np.ma.getmask(new_array)
+            new_data = np.ma.getdata(new_array)
+            new_mask = np.ma.getmask(new_array)
         else:
-            new_map.data = new_array
-            new_map.mask = None
+            new_data = new_array
+            new_mask = None
+
+        #Create new map with the modified data
+        new_map = self._new_instance(new_data, new_meta, self.plot_settings, mask=new_mask)
         return new_map
 
 # #### Visualization #### #
@@ -1526,7 +1557,7 @@ scale:\t\t {scale}
             axes = plt.Axes(figure, [0., 0., 1., 1.])
             axes.set_axis_off()
             figure.add_axes(axes)
-            matplot_args.update({'annotate':False})
+            matplot_args.update({'annotate': False, "_basic_plot": True})
 
         # Normal plot
         else:
@@ -1582,20 +1613,24 @@ scale:\t\t {scale}
 
         """
 
+        # extract hiddden kwarg
+        _basic_plot = imshow_kwargs.pop("_basic_plot", False)
+
         # Get current axes
         if not axes:
             axes = wcsaxes_compat.gca_wcs(self.wcs)
 
-        # Check that the image is properly oriented
-        if (not wcsaxes_compat.is_wcsaxes(axes) and
-            not np.array_equal(self.rotation_matrix, np.matrix(np.identity(2)))):
-            warnings.warn("This map is not properly oriented. Plot axes may be incorrect",
-                          Warning)
+        if not _basic_plot:
+            # Check that the image is properly oriented
+            if (not wcsaxes_compat.is_wcsaxes(axes) and
+                not np.array_equal(self.rotation_matrix, np.matrix(np.identity(2)))):
+                warnings.warn("This map is not properly oriented. Plot axes may be incorrect",
+                              Warning)
 
-        if not wcsaxes_compat.is_wcsaxes(axes) and wcsaxes_compat.HAVE_WCSAXES:
-            warnings.warn("WCSAxes is installed but not being used."
-                          " Plots may not have the expected behaviour.",
-                          Warning)
+            if not wcsaxes_compat.is_wcsaxes(axes) and wcsaxes_compat.HAVE_WCSAXES:
+                warnings.warn("WCSAxes is installed but not being used."
+                              " Plots may not have the expected behaviour.",
+                              Warning)
 
         # Normal plot
         imshow_args = deepcopy(self.plot_settings)
@@ -1638,7 +1673,7 @@ scale:\t\t {scale}
         if wcsaxes_compat.is_wcsaxes(axes):
             wcsaxes_compat.default_wcs_grid(axes)
 
-        #Set current image (makes colorbar work)
+        # Set current image (makes colorbar work)
         plt.sca(axes)
         plt.sci(ret)
         return ret