+++ sunpy/tests/setup_command.py
@@ -109,7 +109,6 @@ class SunPyTest(AstropyTest):
                'online={online!r}, '
                'offline={offline!r}, '
                'figure={figure!r}, '
-               'coverage={1.coverage!r}, '
                'cov_report={1.cov_report!r})); '
                '{cmd_post}'
                'sys.exit(result)')