+++ sunpy/util/tests/test_cond_dispatch.py
@@ -8,7 +8,8 @@ import pytest
 from sunpy.util.cond_dispatch import ConditionalDispatch
 
 
-def pytest_funcarg__oddeven(request):
+@pytest.fixture
+def oddeven(request):
     f = ConditionalDispatch()
     # Multiply even numbers by two.
     f.add(lambda x: 2 * x, lambda x: x % 2 == 0)
@@ -25,7 +26,8 @@ def test_dispatch(oddeven):
 def test_wrong_sig(oddeven):
     with pytest.raises(TypeError) as exc_info:
         oddeven(y=2)
-    assert "There are no functions matching your input parameter signature." in str(exc_info.value)
+    assert "There are no functions matching your input parameter signature." in str(
+        exc_info.value)
 
 
 def test_nocond():
@@ -34,7 +36,8 @@ def test_nocond():
     f.add(lambda x: 2 * x, lambda x: x % 2 == 0)
     with pytest.raises(TypeError) as exc_info:
         f(3)
-        assert "Your input did not fulfill the condition for any function." in str(exc_info.value)
+        assert "Your input did not fulfill the condition for any function." in str(
+            exc_info.value)
 
 
 def test_else():